backend/requirements.txt
# local package
-e .

psycopg2==2.8.2
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
Flask-Migrate==2.4.0
Flask-Dance==2.0.0
gunicorn==19.9.0
flask-restx==0.2.0
dotmap==1.3.8

#######twitter#######
python-twitter==3.5

#######last fm#######
pylast==3.1.0
Pillow==6.2.0

#dev
pylint-flask #handles pylint errors for models
#testing
pytest
pytest-flask